+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - enhancements
 - bugs
+    + Checks for use_pagination method on filter action
     + Fixes filter form variable
 
 ## 1.4.1